Rename `reactive.html` to `reactive.event`
