+++ DevToolkt/reactiveDom/src/jsMain/kotlin/dev/toolkt/dom/reactive/ReactiveDivElement.kt
@@ -2,20 +2,20 @@ package dev.toolkt.dom.reactive
 
 import dev.toolkt.reactive.reactive_list.ReactiveList
 import dev.toolkt.reactive.event_stream.EventEmitter
-import dev.toolkt.dom.reactive.html.HtmlEventHandler
-import dev.toolkt.dom.reactive.html.HtmlMouseEvent
-import dev.toolkt.dom.reactive.html.attach
+import dev.toolkt.dom.reactive.event.ReactiveEventHandler
+import dev.toolkt.dom.reactive.event.ReactiveMouseEvent
+import dev.toolkt.dom.reactive.event.attach
 import org.w3c.dom.events.EventTarget
 
 class ReactiveDivElement(
     override val children: ReactiveList<ReactiveNode>,
-    private val handleMouseDown: HtmlEventHandler<HtmlMouseEvent> = HtmlEventHandler.Accepting,
+    private val handleMouseDown: ReactiveEventHandler<ReactiveMouseEvent> = ReactiveEventHandler.Accepting,
 ) : ReactiveGenericHtmlElement() {
     override val elementName: String = "div"
 
-    private val onMouseDownEmitter = EventEmitter<HtmlMouseEvent>()
+    private val onMouseDownEmitter = EventEmitter<ReactiveMouseEvent>()
 
-    val onMouseDown: EventEmitter<HtmlMouseEvent>
+    val onMouseDown: EventEmitter<ReactiveMouseEvent>
         get() = onMouseDownEmitter
 
     override fun attachEventHandlers(
@@ -24,7 +24,7 @@ class ReactiveDivElement(
         handleMouseDown.attach(
             target = target,
             eventName = "mouseDown",
-            wrapper = HtmlMouseEvent.Companion,
+            wrapper = ReactiveMouseEvent.Companion,
             emitter = onMouseDownEmitter,
         )
     }
