+++ DevToolkt/reactiveDom/src/jsMain/kotlin/dev/toolkt/dom/reactive/event/ReactiveEvent.kt
@@ -1,0 +1,9 @@
+package dev.toolkt.dom.reactive.event
+
+import org.w3c.dom.events.Event
+
+sealed class ReactiveEvent {
+    interface Wrapper<E : ReactiveEvent> {
+        fun wrap(rawEvent: Event): E
+    }
+}
