+++ DevToolkt/reactiveDom/src/jsMain/kotlin/dev/toolkt/dom/reactive/event/ReactiveEventHandler.kt
@@ -1,11 +1,11 @@
-package dev.toolkt.dom.reactive.html
+package dev.toolkt.dom.reactive.event
 
 import dev.toolkt.reactive.event_stream.EventEmitter
 import org.w3c.dom.events.EventTarget
 
-interface HtmlEventHandler<in E : HtmlEvent> {
-    object Accepting : HtmlEventHandler<HtmlEvent> {
-        override fun handle(event: HtmlEvent): Resolution = Resolution.Accept
+interface ReactiveEventHandler<in E : ReactiveEvent> {
+    object Accepting : ReactiveEventHandler<ReactiveEvent> {
+        override fun handle(event: ReactiveEvent): Resolution = Resolution.Accept
     }
 
     enum class Resolution {
@@ -15,10 +15,10 @@ interface HtmlEventHandler<in E : HtmlEvent> {
     fun handle(event: E): Resolution
 }
 
-fun <E : HtmlEvent> HtmlEventHandler<E>.attach(
+fun <E : ReactiveEvent> ReactiveEventHandler<E>.attach(
     target: EventTarget,
     eventName: String,
-    wrapper: HtmlEvent.Wrapper<E>,
+    wrapper: ReactiveEvent.Wrapper<E>,
     emitter: EventEmitter<E>,
 ) {
     target.addEventListener(
@@ -27,7 +27,7 @@ fun <E : HtmlEvent> HtmlEventHandler<E>.attach(
             val wrappedEvent = wrapper.wrap(rawEvent = rawEvent)
             val resolution = handle(wrappedEvent)
 
-            if (resolution == HtmlEventHandler.Resolution.PreventDefault) {
+            if (resolution == ReactiveEventHandler.Resolution.PreventDefault) {
                 rawEvent.preventDefault()
             } else {
                 emitter.emit(wrappedEvent)
