--- DevToolkt/reactiveDom/src/jsMain/kotlin/dev/toolkt/dom/reactive/html/HtmlEvent.kt
@@ -1,9 +1,0 @@
-package dev.toolkt.dom.reactive.html
-
-import org.w3c.dom.events.Event
-
-sealed class HtmlEvent {
-    interface Wrapper<E : HtmlEvent> {
-        fun wrap(rawEvent: Event): E
-    }
-}
